refactor(decisions): extract override counting into its own module (BI-ACF0D6D4)
`evaluator.ts` crossed the 800-LOC ceiling again at 841 — the override-count
change carries a long rationale comment, because "what counts as the owner
having overruled us" is a governance judgement rather than a mechanical detail.

That judgement is worth reading on its own instead of buried mid-evaluator, so
`getRecentOverrideCount` moves to `override-count.ts` with the window constant
it depends on. `RECENT_OVERRIDE_WINDOW_DAYS` is re-exported from the evaluator
for existing importers, with the new module as its single source of truth
rather than a second copy.

evaluator.ts 841 -> 782 LOC; new module 70. Pure extraction — no logic
changed, and the suite that covers it targets the real function directly.

1906 tests green; module-size guard clean.

Design-Grounding-Decision: no-contract-change (extraction to satisfy the module-size ceiling; behaviour pinned by the existing suite)
